requirements.txt
fastapi
uvicorn
fpdf==1.7.2
jinja2
python-multipart
pydantic
# openai-whisper
itsdangerous
supabase==1.0.3
python-dotenv
requests
psycopg2-binary
